Moving some functions to the utils file.
