+++ src/app/utils/utils.js
@@ -1,4 +1,8 @@
-import { union as _union } from 'underscore';
+import {
+  union as _union,
+  contains as _contains,
+  each as _each,
+} from 'underscore';
 
 import appConfig from '../../../appConfig.js';
 import { gaUtils } from 'dgx-react-ga';
@@ -17,6 +21,38 @@ function Utils() {
    * @param {string} label Label for GA event.
    */
   this.trackPicks = gaUtils.trackEvent('Staff Picks');
+
+  /**
+   * getPickTags(book)
+   * Return an array of the pick's tags, lowercased and with a hyphen to easily
+   * use as a class or an ID.
+   *
+   * @param {object} book
+   */
+  this.getPickTags = (book) => {
+    if (!(book.tags && book.tags.length)) {
+      return [];
+    }
+    return book.tags.map(tag => tag.toLowerCase().split(' ').join('-'));
+  };
+
+  /**
+   * getSelectedTags(tagArray, selectedFilters)
+   * Return an array of the filters in a pick's tagArray if the selectedFilters is found.
+   *
+   * @param {array} tagArray
+   * @param {array} selectedFilters
+   */
+  this.getSelectedTags = (tagArray, selectedFilters) => {
+    const inSelectedFilter = [];
+    _each(tagArray, (bookTag) => {
+      if (_contains(selectedFilters, bookTag)) {
+        inSelectedFilter.push(bookTag);
+      }
+    });
+
+    return inSelectedFilter;
+  };
 }
 
 export default new Utils();
